% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCap_API.R
\name{build_rcon}
\alias{build_rcon}
\title{Build REDCap API Connection}
\usage{
build_rcon(token, url = "https://redcap.med.uvm.edu/api/")
}
\arguments{
\item{token}{REDCap-generated unique API key.}

\item{url}{https://redcap.med.uvm.edu/api/}
}
\value{
REDCap Connection Object
}
\description{
Build REDCap API Connection
}
\examples{
\dontrun{
token <- read.table("my_secret_key.txt")
build_rcon(token)
}
}
