% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivr.R
\name{clean_ivr}
\alias{clean_ivr}
\title{Clean IVR Data}
\usage{
clean_ivr(ivr_raw, enrl)
}
\arguments{
\item{ivr_raw}{Exported from tcoRs::load_ivr()}

\item{enrl}{Exported from tcoRs::get_maintrial_enrollment()}
}
\value{
Data frame: one row per subject per day.
}
\description{
Clean IVR Data
}
\examples{
\dontrun{
ivr_raw <- tcoRs::load_ivr(token = ivr_token)
enrl <- tcoRs::get_maintrial_enrollment(token = REDCap_token)
ivr_clean <- clean_ivr(ivr_raw, enrl)
}
}
