% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adverse_events.R
\name{count_ae}
\alias{count_ae}
\title{Count Adverse Events}
\usage{
count_ae(ae_clean, var, by_condition = FALSE)
}
\arguments{
\item{ae_clean}{exported from `tcoRs::summarize_ae()`}

\item{var}{quoted string of summary variable, e.g. "ae_sev"}

\item{by_condition}{summarize output by treatment group; default = `FALSE`}
}
\value{
vector of data frames, subject count by group, and AE-subject counts by group
}
\description{
count distinct subject-variable combinations,
}
\examples{
\dontrun{
count_ae(ae_clean, "ae_sev")
}
}
