% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ctcae5}
\alias{ctcae5}
\title{NIH National Cancer Institute Common Terminology Criteria for Adverse Events (CTCAE)}
\format{
An object of class \code{data.frame} with 837 rows and 11 columns.
}
\source{
\url{https://ctep.cancer.gov/protocoldevelopment/electronic_applications/ctc.htm#ctc_60}
}
\usage{
ctcae5
}
\description{
A crosswalk data set containing common terminologies for adverse events and
}
\keyword{datasets}
