% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCap_API.R
\name{download_rc_dataframe}
\alias{download_rc_dataframe}
\title{Download REDCap data frame, a wrapper around `exportRecords` with helpful defaults.}
\usage{
download_rc_dataframe(rcon, fields = NULL, events = NULL)
}
\arguments{
\item{rcon}{REDCap API connection (see `build_rcon()`).}

\item{fields}{Character vector of field names.}

\item{events}{Character vector of event names.}
}
\value{
data frame
}
\description{
Download REDCap data frame, a wrapper around `exportRecords` with helpful defaults.
}
\examples{
\dontrun{
rcon <- build_rcon(token)
download_rc_dataframe(rcon, fields = "screen_id", events = "screening_arm_1")
}
}
