% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrol-data.R
\docType{data}
\name{enrl_sample}
\alias{enrl_sample}
\title{Sample Main Trial Enrollment data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 8 columns.
}
\source{
Main Trial REDCap Project
}
\usage{
data(enrl_sample)
}
\description{
Session log export used to demonstrate helper functions.
}
\keyword{datasets}
