% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmb.R
\name{get_dsmb_measures}
\alias{get_dsmb_measures}
\title{Get DSMB REDCap Measures}
\usage{
get_dsmb_measures(rcon)
}
\arguments{
\item{rcon}{REDCap connection object exported from tcoRS::build_rcon()}
}
\value{
Data frame of longitudinal trial measures grouped into weeks
}
\description{
Get DSMB REDCap Measures
}
\examples{
\dontrun{
rcon <- build_rcon("my_secret_token")
df <- get_dsmb_measures(rcon)
}
}
