% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adverse_events.R
\name{get_maintrial_aes}
\alias{get_maintrial_aes}
\title{Get Main Trial Adverse Events}
\usage{
get_maintrial_aes(rcon, fields = field_vectors$ae_fields)
}
\arguments{
\item{rcon}{REDCap connection object exported from tcoRS::build_rcon()}

\item{fields}{pre-determined vector of adverse event fields}
}
\value{
data frame of adverse events
}
\description{
Get Main Trial Adverse Events
}
\examples{
\dontrun{
rcon <- build_rcon("my_secret_token")
ae_df <- get_maintrial_aes(rcon)
}
}
