% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_maintrial_conditions.R
\name{get_maintrial_conditions}
\alias{get_maintrial_conditions}
\title{Get Main Trial Treatment Groups}
\usage{
get_maintrial_conditions(username, password)
}
\arguments{
\item{username}{UMN Account Username}

\item{password}{UMN Account Password}
}
\value{
Data Frame with condition assignments
}
\description{
Get official dose conditions from randomization database. This database is hosted
at the University of Minnesota.
}
\examples{
\dontrun{
conditions <- get_maintrial_conditions(token = "my_secret_token")
}
}
