% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrollment.R
\name{get_maintrial_enrollment}
\alias{get_maintrial_enrollment}
\title{Pull Main Trial Enrollment Information}
\usage{
get_maintrial_enrollment(rcon, fields = field_vectors$enrollment_fields)
}
\arguments{
\item{rcon}{REDCap connection object exported from tcoRs::build_rcon()}

\item{fields}{A selection of pre-determined enrollment REDCap fields}
}
\value{
Data frame, one row per subject
}
\description{
Pull Main Trial Enrollment Information
}
\examples{
\dontrun{
enrl <- get_maintrial_enrollment(rcon)
}
}
