% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrollment.R
\name{get_maintrial_prescreens}
\alias{get_maintrial_prescreens}
\title{Pull Main Trial Prescreens}
\usage{
get_maintrial_prescreens(
  rcon,
  site_name,
  min_date = as.Date("2020-10-01"),
  fields = field_vectors$ps_inel_names
)
}
\arguments{
\item{rcon}{REDCap Connection object exported from tcoRs::build_rcon()}

\item{site_name}{String name of study site, appended to return.}

\item{min_date}{Minimum prescreen date, fixed to 2020-10-01.}

\item{fields}{Vector of pre-determined REDCap prescreen eligibility fields.}
}
\value{
Data frame, one row per prescreen.
}
\description{
Pull Main Trial Prescreens
}
\examples{
\dontrun{
rcon_uvm <- build_rcon("my_secret_token")
get_maintrial_prescreens <- (rcon = rcon_uvm, site_name = "uvm")
}
}
