% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{is.redcapFactor}
\alias{is.redcapFactor}
\title{REDCap Factors}
\usage{
is.redcapFactor(v)
}
\arguments{
\item{v}{a vector of data containing class `redcapFactor`}
}
\value{
`TRUE` or `FALSE`
}
\description{
REDCap Factors
}
\examples{
\dontrun{
is.redcapFactor(df$rc_factor_var)
}
}
