% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmb.R
\name{ivr_count}
\alias{ivr_count}
\title{Number of Participants by Week Bin}
\usage{
ivr_count(ivr_sum)
}
\arguments{
\item{ivr_sum}{exported from tcoRs::summarize_ivr()}
}
\value{
tally of participants (1) by group, (2) by binned week
}
\description{
Number of Participants by Week Bin
}
\examples{
\dontrun{
ivr_count(ivr_sum)
}
}
