% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmb.R
\name{ivr_ecig_days}
\alias{ivr_ecig_days}
\title{IVR Ecig Days}
\usage{
ivr_ecig_days(ivr_sum)
}
\arguments{
\item{ivr_sum}{Summary IVR data frame}
}
\value{
Vector of average e-cig use-days per week, and special quantities
}
\description{
IVR Ecig Days
}
\details{
Number of days per week participants used ecigarettes
}
\examples{
\dontrun{
ivr_ecig_days(ivr_sum)
}
}
