% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmb.R
\name{ivr_summary_dsmb}
\alias{ivr_summary_dsmb}
\title{IVR Use Summary DSMB}
\usage{
ivr_summary_dsmb(ivr_sum)
}
\arguments{
\item{ivr_sum}{summarized IVR data frame from tcoRs::summarize_ivr()}
}
\value{
vector of cig and ecig use
}
\description{
IVR Use Summary DSMB
}
\examples{
\dontrun{
ivr_summary_dsmb(ivr_sum)
}
}
