% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmb.R
\name{ivr_weekbin_dsmb}
\alias{ivr_weekbin_dsmb}
\title{IVR Weekbin DSMB}
\usage{
ivr_weekbin_dsmb(ivr_sum)
}
\arguments{
\item{ivr_sum}{summarized IVR data frame from tcoRs::summarize_ivr()}
}
\value{
data frame with `week_bin` column
}
\description{
Impose week bins on IVR data for main trial DSMB summary
}
\examples{
\dontrun{
ivr_weekbin_dsmb(ivr_sum)
}
}
