% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivr.R
\name{load_ivr}
\alias{load_ivr}
\title{Download Main Trial IVR Data}
\usage{
load_ivr(token)
}
\arguments{
\item{token}{TeleSage Raw Data Download key}
}
\value{
Data frame of all surveys.
}
\description{
Download Main Trial IVR Data
}
\examples{
\dontrun{
df <- load_ivr(token)
}
}
