% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{matrix_to_vector}
\alias{matrix_to_vector}
\title{Convert matrix data to character vector}
\usage{
matrix_to_vector(df, colname, matrix_vars, select_string)
}
\arguments{
\item{df}{Data frame containing (not necessarily exclusively) matrix data.}

\item{colname}{String of resulting column name}

\item{matrix_vars}{Named list of matrix variables.}

\item{select_string}{String to uniquely identify matrix variables to remove using dplyr::contains()}
}
\value{
Data frame with vectorized matrix variables.
}
\description{
Convert matrix data to character vector
}
\examples{
\dontrun{
df <- matrix_to_vector(df, "new_col", matrix_list, "matrix_string")
}
}
