% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmb.R
\name{odd_even_dsmb}
\alias{odd_even_dsmb}
\title{Odd Even DSMB}
\usage{
odd_even_dsmb(df)
}
\arguments{
\item{df}{REDCap Data Frame}
}
\value{
vector of data frames, (1) odd weeks, (2) even weeks
}
\description{
Odd Even DSMB
}
\examples{
\dontrun{
df_vector <- odd_evn_dsmb(df)
}
}
