% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{pi_prop}
\alias{pi_prop}
\title{Impose Pilot and Proper}
\usage{
pi_prop(df)
}
\arguments{
\item{df}{Data frame with `screen_id` variable of the form `X-A###`}
}
\value{
Data frame with column appended `pi_prop` to designate study proper from pilot.
}
\description{
Impose Pilot and Proper
}
\examples{
\dontrun{
df <- pi_prop(df)
}
}
