% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmb.R
\name{pivot_week_summary}
\alias{pivot_week_summary}
\title{Pivot Week Summary}
\usage{
pivot_week_summary(df)
}
\arguments{
\item{df}{Summarized data frame}
}
\value{
data frame
}
\description{
Pivot Week Summary
}
\examples{
\dontrun{
pivot_week_summary(df_vector[[1]])
}
}
