% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{pjt_ste}
\alias{pjt_ste}
\title{Impose Project and Site}
\usage{
pjt_ste(df)
}
\arguments{
\item{df}{Data frame with `screen_id` variable of the form `X-A###`}
}
\value{
Data frame with columns appended for `project` and `site`
}
\description{
Impose Project and Site
}
\examples{
\dontrun{
df <- pjt_ste(df)
}
}
