% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmb.R
\name{protocol_violations}
\alias{protocol_violations}
\title{Protocol Violations DSMB}
\usage{
protocol_violations(rcon, cond)
}
\arguments{
\item{rcon}{REDCap connection object exported from tcoRs::build_rcon()}

\item{cond}{data frame of condition assignments from tcoRs::get_maintrial_conditions()}
}
\value{
data frame of protocol violations
}
\description{
Protocol Violations DSMB
}
\examples{
\dontrun{
rcon <- build_rcon("my_secret_token")
cond <- get_maintrial_conditions(usr, pws)
violations <- protocol_violations(rcon, cond)
}
}
