% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmb.R
\name{randomization_accrual_plot}
\alias{randomization_accrual_plot}
\title{Plot randomization progress}
\usage{
randomization_accrual_plot(enrl, goal, min_date, end_date, proj, title)
}
\arguments{
\item{enrl}{data frame exported from tcoRs::get_maintrial_enrollment()}

\item{goal}{numeric constant of project's randomization goal}

\item{min_date}{string enrollment start date, "yyyy-mm-dd"}

\item{end_date}{string projected enrollment end date, "yyyy-mm-dd"}

\item{proj}{string of either "Project 1", "Project 2", or "Project 3"}

\item{title}{string, title of plot}
}
\value{
ggplot2 object
}
\description{
Plot randomization progress
}
\examples{
\dontrun{
randomization_accrual_plot(
enrl,
goal = 212,
min_date = "2020-11-01",
end_date = "2023-06-01",
proj = "Project 1",
title = "Recruitment Accrual Graph: Project 1"

)
}
}
