% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrollment.R
\name{session_distribution}
\alias{session_distribution}
\title{Session Distribution}
\usage{
session_distribution(rcon)
}
\arguments{
\item{rcon}{REDCap connection object exported from `tcors::build_rcon()`}
}
\value{
data frame of session counts
}
\description{
returns a tally of sessions completed for each project
}
\examples{
\dontrun{
session_distribution(rcon)
}
}
