% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adverse_events.R
\name{summarize_ae}
\alias{summarize_ae}
\title{Summarize AE}
\usage{
summarize_ae(ae_clean, cond, filter_proper = TRUE)
}
\arguments{
\item{ae_clean}{export from `tcoRs::clean_ae()`}

\item{cond}{export from `tcoRs::get_maintrial_conditions()`}

\item{filter_proper}{filter for study proper IDs; default = `TRUE`}
}
\value{
vector of clean AEs and a set to review
}
\description{
Summarize AE
}
\examples{
\dontrun{
summarize_ae(ae_clean, cond)
}
}
