% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrollment.R
\name{summarize_demographics}
\alias{summarize_demographics}
\title{Summarize Demographics}
\usage{
summarize_demographics(dem, enrl, which_table)
}
\arguments{
\item{dem}{data frame exported from tcoRs::demographics()}

\item{enrl}{data frame exorted from tcoRS::get_maintrial_conditions()}

\item{which_table}{one of
* "table1" : characteristics of all screenings by site
* "table2" : characteristics of randomized participants by site
* "table3" : characteristics of all randomized partiicpants by treatment grouop}
}
\value{
3 vectors
}
\description{
Summarize Demographics
}
\examples{
\dontrun{
vector <- summarize_demographics(dem, enrl, "table1")
}
}
