% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmb.R
\name{summarize_evali}
\alias{summarize_evali}
\title{Summarize EVALI symptoms}
\usage{
summarize_evali(df)
}
\arguments{
\item{df}{data frame exported from tcoRs::get_dsmb_measures()}
}
\value{
data frame of EVALI severe symtpoms summarized
}
\description{
Summarize EVALI symptoms
}
\examples{
\dontrun{
summarize_evali(redcap_values)
}
}
