% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivr.R
\name{summarize_ivr}
\alias{summarize_ivr}
\title{Summarize IVR Data Set}
\usage{
summarize_ivr(ivr_clean, conditions)
}
\arguments{
\item{ivr_clean}{Exported from tcoRs::clean_ivr()}

\item{conditions}{Exported from tcoRs::get_maintrial_conditions()}
}
\value{
Data Frame -- one entry for each subject for each week.
}
\description{
Obtain weekly summary values for each subject from daily surveys.
}
\examples{
\dontrun{
ivr_sum <- summarize_ivr(ivr_clean, conditions)
}
}
