% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrollment.R
\name{summarize_maintrial_prescreens}
\alias{summarize_maintrial_prescreens}
\title{Summarize Main Trial Prescreens}
\usage{
summarize_maintrial_prescreens(df)
}
\arguments{
\item{df}{Data Frame exported from tcoRs::get_maintrial_prescreens()}
}
\value{
List of data frames with (1) prsecreen tallies by site and (2) by exclusion criterion
}
\description{
Summarize Main Trial Prescreens
}
\examples{
\dontrun{
ps_vector <- summarize_maintrial_prescreens(ps_df)
}
}
