% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrollment.R
\name{summarize_maintrial_screenings}
\alias{summarize_maintrial_screenings}
\title{Summarize Maintrial Screenings}
\usage{
summarize_maintrial_screenings(df_enrl, filter_proper = TRUE)
}
\arguments{
\item{df_enrl}{Data frame exported from tcoRs::get_maintrial_enrollment()}

\item{filter_proper}{Filter only study proper IDs. If `FALSE`, all IDs are counted.}
}
\value{
List of data frames with
1. Screening tallies by site
2. Exclusion criteria
}
\description{
Summarize Maintrial Screenings
}
\examples{
\dontrun{
scrn_vector <- summarize_maintrial_screenings(df_enrl)
}
}
