% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmb.R
\name{value_summary_dsmb}
\alias{value_summary_dsmb}
\title{Value Summary DSMB}
\usage{
value_summary_dsmb(df, variable)
}
\arguments{
\item{df}{data frame with value, week bin, condition}

\item{variable}{string, one of "co", "bdi", or "oasis"}
}
\value{
data frame of values, data frame of subject counts
}
\description{
Summarize single-column quantity with week-bin variable
}
\examples{
\dontrun{
df <- tcoRs::get_dsmb_measures(rcon) \%>\%
dplyr::left_join(cond, by = "screen_id") \%>\%
tcoRs::pjt_ste() \%>\%
value_summary_dsmb("co")
}
}
