% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmb.R
\name{week_bin_dsmb}
\alias{week_bin_dsmb}
\title{DSMB: Week Bin}
\usage{
week_bin_dsmb(df)
}
\arguments{
\item{df}{REDCap Data Frame}
}
\value{
data frame with `session` in binned weeks appended
}
\description{
DSMB: Week Bin
}
\examples{
\dontrun{
df <- week_bin_dsmb(df)
}
}
